AttackCoolcown V0.06
The closest I'm going to be able to get NPCs not to spam attack. If they do get to spam strike a player then it's much more rare and just a bug within the mod itself. May attempt to lurk more into it.

For now, the next task is to fix the bug where the NPC gets hit once, they run away. Not all npc's should do this. Especially with raiders.