NewEntitySpawner V0.01
First edition of making NPC's spawn. I've noticed the other form of NPC's spawning seem to be more dumb, but not the raiders. The raiders seem to have a bit more INT. So copying the Raider's Data and just make this version 'not hostile' allows the world to spawn random npcs in the world. It's very primitive, and will need in game settings editing for spawn adjustment, but it works! npcs that spawn will pick a home and hold up there.
Maybe an idea here would be giving the npcs hostility to anyone that goes in their base from other npcs. The raiders do a pretty good job at this so far.