AttackCooldown V0.07
This version adjusts the cooldowns. Now that there is a bigger tickcooldown on the refresh rate of checking npcs actions, the AtkTicks are 'twice as slow'. Which isn't BAD, just now the cooldown timers act almost like real life seconds.

Also found a bug where "n = 1.0f" shouldn't exist in a random line of code. Must of been an accident from previous dev. I double checked to be sure, there wasn't meant to be an 'f' there in the result code.

Also adjusted a few of the 'if stuck ticks' to be smaller so that (in an attempt) to make the npcs not stay in place as long as previously. This code may not be useful much longer since I'm already working on the other part of the code that makes npcs not get stuck as often.