AI-Manager Update General
General code update/cleanup to be easier to read.
Did make an attempt to fix the "When raider decides to make a new base" code, and to check for a few more things before making it.
Still need to check into seeing if they stop their task half way through when something is in the same building as them